delivery_point advances

+++ Lab2TBD/Backend/src/main/java/com/example/Lab1TBD/persistence/entities/EstablishmentEntity.java
@@ -8,7 +8,7 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class EstablishmentEntity {
-    private Integer establishment_id;   //identificador del establecimiento
+    private Long establishment_id;   //identificador del establecimiento
     private String establishment_data;  //data del establecimiento
     private String region_data;         //region del establecimiento
     private String latitude;            //latitud del establecimiento
